requirements.txt
Django==1.8.3
Markdown==2.6.2
Pillow==2.9.0
PyMySQL==0.6.6
Pygments==2.0.2
defusedxml==0.4.1
django-allauth==0.22.0
django-appconf==1.0.1
django-compressor==1.5
django-filter==0.10.0
djangorestframework==3.1.3
httpie==0.9.2
oauthlib==1.0.1
python-dateutil==2.4.2
python3-openid==3.0.5
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
